Fix compile issues
